refactor: move resources to own component
