+++ src/components/RelatedContentModules/Resources.js
@@ -1,0 +1,116 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { css } from '@emotion/core';
+import { graphql, Link } from 'gatsby';
+import { ExternalLink, Icon } from '@newrelic/gatsby-theme-newrelic';
+import Section from './Section';
+import Tag from '../Tag';
+import Title from './Title';
+
+const SITE_TAGS = {
+  developer: 'https://developer.newrelic.com',
+  'open source': 'https://opensource.newrelic.com',
+  docs: 'https://docs.newrelic.com',
+};
+
+const Resources = ({ page }) => {
+  const {
+    frontmatter: { resources },
+  } = page;
+
+  return resources?.length > 0 ? (
+    <Section>
+      <Title>Resources</Title>
+      <nav>
+        <ul
+          css={css`
+            list-style: none;
+            margin: 0;
+            padding: 0;
+          `}
+        >
+          {resources.map((resource) => {
+            const tag = resource.url.startsWith('/')
+              ? 'developer'
+              : Object.keys(SITE_TAGS).find((tag) =>
+                  resource.url.startsWith(SITE_TAGS[tag])
+                );
+
+            const isDeveloperSite = tag === 'developer';
+
+            const LinkElement = isDeveloperSite ? Link : ExternalLink;
+            const props = isDeveloperSite
+              ? {
+                  to: resource.url.replace(
+                    'https://developer.newrelic.com',
+                    ''
+                  ),
+                }
+              : { href: resource.url };
+
+            return (
+              <li
+                key={resource.url}
+                css={css`
+                  margin-bottom: 1rem;
+                `}
+              >
+                <LinkElement
+                  {...props}
+                  css={css`
+                    display: block;
+                    margin-bottom: 0.25rem;
+                  `}
+                >
+                  <span
+                    css={css`
+                      vertical-align: middle;
+                    `}
+                  >
+                    {resource.title}
+                  </span>
+                  {!isDeveloperSite && (
+                    <Icon
+                      name={Icon.TYPE.EXTERNAL_LINK}
+                      css={css`
+                        margin-left: 0.25rem;
+                        vertical-align: middle;
+                      `}
+                    />
+                  )}
+                </LinkElement>
+                <Tag>{tag}</Tag>
+              </li>
+            );
+          })}
+        </ul>
+      </nav>
+    </Section>
+  ) : null;
+};
+
+Resources.propTypes = {
+  page: PropTypes.shape({
+    frontmatter: PropTypes.shape({
+      resources: PropTypes.arrayOf(
+        PropTypes.shape({
+          title: PropTypes.string,
+          url: PropTypes.string,
+        })
+      ),
+    }),
+  }),
+};
+
+export const query = graphql`
+  fragment Resources_page on Mdx {
+    frontmatter {
+      resources {
+        title
+        url
+      }
+    }
+  }
+`;
+
+export default Resources;
