+++ src/components/RelatedContentModules/index.js
@@ -1,2 +1,3 @@
 export { default as Contribute } from './Contribute';
 export { default as PageUpdated } from './PageUpdated';
+export { default as Resources } from './Resources';
